% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc.R
\name{escServer}
\alias{escServer}
\title{escServer}
\usage{
escServer(id = "esc", i, o, ...)
}
\arguments{
\item{id}{module namespace id "esc"}

\item{i}{shiny input}

\item{o}{shiny output}

\item{...}{any other NAMED objects from main shiny process to pass to browser}
}
\value{
esckey module server
}
\description{
Place \code{escServer(i = input, o = output)} call somewhere within your \code{server} definition.
Respond to \code{esc} keypress with \code{browser()}
Must provide names for objects provided with \code{...} eg. \code{foo = foo}
}
