% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc.R
\name{escUI}
\alias{escUI}
\title{escUI}
\usage{
escUI(id = "esc")
}
\arguments{
\item{id}{"esc"}
}
\value{
esckey module UI
}
\description{
Place \code{escUI()} call somewhere within your \code{ui} definition
}
\examples{
\dontrun{
# Placed in app UI function
escUI()
}
}
